+++ CHANGES.rst
@@ -14,6 +14,11 @@ Changelog
 - Removed `Meeting.items` `ReferenceField`, manage it manually,
   this will help migrating to `DX`.
   [gbastien]
+- Optimized `utils.sendMailIfRelevant` to not send an email several times to
+  same address.  It was only done in `MeetingItem._sendMailToGroupMembers`,
+  now this method is removed and it is managed in `utils.sendMailIfRelevant`
+  using parameter `isGroupIds=True`.
+  [gbastien]
 
 4.2b6 (2020-11-19)
 ------------------